Separate out client-specific Rust code from worker code
